requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
PyJWT==2.5.0
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3